closing ISSSUE: organizing-animation-tools
