+++ bouncingBall/index.html
@@ -6,7 +6,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
     <link rel="stylesheet" href="./style.css">
-    <script src="../animate.js"></script>
 </head>
 <body>
     <h4>Click On the Ball</h4>
@@ -18,5 +17,5 @@
             </svg>
         <div id="floor"></div>
     </div>
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
 </body>
