+++ bouncingBall/script.js
@@ -1,11 +1,13 @@
+import * as animTools from '../tools.js';
+
 let ball = document.getElementById('football');
 
 let _ball = {
     draw: function(progress){
         ball.style.top = `${40 * progress}vh`;
     },
-    timingFunction: timingFunctions.pseudoBounce(.5, 4, 2),
+    timingFunction: animTools.timingFunctions.pseudoBounce(.5, 4, 2),
     duration: 3000
 };
 
-ball.onclick = () => animate(_ball);
+ball.onclick = () => animTools.animate(_ball);
