+++ tools.js
@@ -1,0 +1,3 @@
+export { animate } from './animate.js';
+export { alterTimingFunction } from './alterTimingFunction.js';
+export { timingFunctions } from './timingFunction.js';
